REFACTOR: Separated sub-components into separate files and using folders for grouping
