+++ app/auth/login/_components/buttons/social-auth-buttons.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { BetterAuthActionButton } from "@/components/auth/better-auth-action-button"
+import { BetterAuthActionButton } from "@/components/auth/buttons/better-auth-action-button"
 import { Button } from "@/components/ui/button"
 import { authClient } from "@/lib/auth/auth-client"
 import {
