+++ app/auth/login/_components/forms/email-verification.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { BetterAuthActionButton } from "@/components/auth/better-auth-action-button";
+import { BetterAuthActionButton } from "@/components/auth/buttons/better-auth-action-button";
 import { authClient } from "@/lib/auth/auth-client";
 import { useEffect, useRef, useState } from "react";
