+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import { Toaster } from "@/components/ui/sonner";
-import { ImpersonationIndicator } from "@/components/auth/impersonation-indicator";
+import { ImpersonationIndicator } from "@/components/auth/buttons/impersonation-indicator";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
