+++ app/organizations/_components/tabs/invites-tab.tsx
@@ -1,4 +1,4 @@
-"use client"
+"use client";
 
 import {
   Table,
@@ -7,21 +7,21 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from "@/components/ui/table"
-import { Badge } from "@/components/ui/badge"
-import { authClient } from "@/lib/auth/auth-client"
-import { BetterAuthActionButton } from "@/components/auth/better-auth-action-button"
-import { CreateInviteButton } from "./create-invite-button"
+} from "@/components/ui/table";
+import { Badge } from "@/components/ui/badge";
+import { authClient } from "@/lib/auth/auth-client";
+import { BetterAuthActionButton } from "@/components/auth/buttons/better-auth-action-button";
+import { CreateInviteButton } from "../buttons/create-invite-button";
 
 /**
  * Shows pending invitations and allows cancellation or new invites.
  * @returns Invitations tab populated from the active organization hook.
  */
 export function InvitesTab() {
-  const { data: activeOrganization } = authClient.useActiveOrganization()
+  const { data: activeOrganization } = authClient.useActiveOrganization();
   const pendingInvites = activeOrganization?.invitations?.filter(
-    invite => invite.status === "pending"
-  )
+    (invite) => invite.status === "pending"
+  );
 
   /**
    * Cancels an outstanding organization invitation.
@@ -29,7 +29,7 @@ export function InvitesTab() {
    * @returns Promise for the cancel request.
    */
   function cancelInvitation(invitationId: string) {
-    return authClient.organization.cancelInvitation({ invitationId })
+    return authClient.organization.cancelInvitation({ invitationId });
   }
 
   return (
@@ -48,7 +48,7 @@ export function InvitesTab() {
           </TableRow>
         </TableHeader>
         <TableBody>
-          {pendingInvites?.map(invitation => (
+          {pendingInvites?.map((invitation) => (
             <TableRow key={invitation.id}>
               <TableCell>{invitation.email}</TableCell>
               <TableCell>
@@ -71,5 +71,5 @@ export function InvitesTab() {
         </TableBody>
       </Table>
     </div>
-  )
+  );
 }
