+++ app/organizations/_components/tabs/members-tab.tsx
@@ -10,7 +10,7 @@ import {
 } from "@/components/ui/table"
 import { Badge } from "@/components/ui/badge"
 import { authClient } from "@/lib/auth/auth-client"
-import { BetterAuthActionButton } from "@/components/auth/better-auth-action-button"
+import { BetterAuthActionButton } from "@/components/auth/buttons/better-auth-action-button"
 
 /**
  * Lists active organization members with removal controls.
