+++ app/profile/_components/account/linked-accounts-tab.tsx
@@ -1,0 +1,25 @@
+import { Card, CardContent } from "@/components/ui/card";
+import { auth } from "@/lib/auth/auth";
+import { headers } from "next/headers";
+import { AccountLinking } from "./account-linking";
+
+/**
+ * Server component that lists linked social accounts and masks credential providers.
+ * @returns Card section with account linking controls.
+ */
+export async function LinkedAccountsTab() {
+  const accounts = await auth.api.listUserAccounts({
+    headers: await headers(),
+  });
+  const nonCredentialAccounts = accounts.filter(
+    (a) => a.providerId !== "credential"
+  );
+
+  return (
+    <Card>
+      <CardContent>
+        <AccountLinking currentAccounts={nonCredentialAccounts} />
+      </CardContent>
+    </Card>
+  );
+}
