+++ components/auth/buttons/better-auth-action-button.tsx
@@ -1,7 +1,7 @@
-"use client"
+"use client";
 
-import { ComponentProps } from "react"
-import { ActionButton } from "../ui/action-button"
+import { ComponentProps } from "react";
+import { ActionButton } from "@/components/ui/action-button";
 
 /**
  * Wraps `ActionButton` with Better Auth's error shape and optional success toast.
@@ -14,21 +14,21 @@ export function BetterAuthActionButton({
   successMessage,
   ...props
 }: Omit<ComponentProps<typeof ActionButton>, "action"> & {
-  action: () => Promise<{ error: null | { message?: string } }>
-  successMessage?: string
+  action: () => Promise<{ error: null | { message?: string } }>;
+  successMessage?: string;
 }) {
   return (
     <ActionButton
       {...props}
       action={async () => {
-        const res = await action()
+        const res = await action();
 
         if (res.error) {
-          return { error: true, message: res.error.message || "Action failed" }
+          return { error: true, message: res.error.message || "Action failed" };
         } else {
-          return { error: false, message: successMessage }
+          return { error: false, message: successMessage };
         }
       }}
     />
-  )
+  );
 }
