+++ lib/auth/o-auth-providers.ts
@@ -1,11 +1,11 @@
-import { DiscordIcon, GitHubIcon } from "@/components/auth/o-auth-icons"
-import { ComponentProps, ElementType } from "react"
+import { DiscordIcon, GitHubIcon } from "@/components/auth/icons/o-auth-icons";
+import { ComponentProps, ElementType } from "react";
 
 /**
  * OAuth providers enabled in the demo experience.
  */
-export const SUPPORTED_OAUTH_PROVIDERS = ["github", "discord"] as const
-export type SupportedOAuthProvider = (typeof SUPPORTED_OAUTH_PROVIDERS)[number]
+export const SUPPORTED_OAUTH_PROVIDERS = ["github", "discord"] as const;
+export type SupportedOAuthProvider = (typeof SUPPORTED_OAUTH_PROVIDERS)[number];
 
 /**
  * Provider metadata used to render provider-specific UI affordances.
@@ -16,4 +16,4 @@ export const SUPPORTED_OAUTH_PROVIDER_DETAILS: Record<
 > = {
   discord: { name: "Discord", Icon: DiscordIcon },
   github: { name: "GitHub", Icon: GitHubIcon },
-}
+};
